+++ src/RabbitMQ.EventBus.AspNetCore/Factories/IRabbitMQPersistentConnection.cs
@@ -31,6 +31,5 @@ namespace RabbitMQ.EventBus.AspNetCore.Factories
         /// </summary>
         /// <returns></returns>
         IModel CreateModel();
-
     }
 }
